requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
Jinja2==3.1.2
Werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
